+++ nodejs-basics/src/modules/esm.mjs
@@ -1,0 +1,38 @@
+import path from 'path';
+import { release, version } from 'os';
+import { createServer } from 'http';
+import { fileURLToPath } from 'url';
+import './files/c.js';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+
+const random = Math.random();
+const targetFile = random > 0.5 ? './files/a.json' : './files/b.json';
+
+const options = { assert: { type: 'json' } };
+const { default: unknown } = await import(targetFile, options);
+
+console.log(unknown);
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const myServer = createServer((_, res) => {
+  res.end('Request accepted');
+});
+
+const PORT = 3000;
+
+console.log(unknown);
+
+myServer.listen(PORT, () => {
+  console.log(`Server is listening on port ${PORT}`);
+  console.log('To terminate it, use Ctrl+C combination');
+});
+
+export { unknown, myServer };
